+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVY014WMTG2Q88V9QNJNT2GJ"
+        "01JVY60BJVANFFTP8M2WJ7P7CB"
     )
     r"""The id of the resource"""